isdin_permission/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
db-dtypes==1.1.1
decorator==5.1.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2023.10.0
gcsfs==2023.10.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.10.2
google-auth==2.23.3
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.12.0
google-cloud-core==2.4.1
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.48.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-oauth==0.14.1
idna==3.7
importlib-metadata==6.11.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
meteostat==1.6.7
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==2.2.2
patsy==0.5.6
Pillow==9.5.0
proto-plus==1.23.0
protobuf==3.20.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.0.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.13.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
statsmodels==0.14.0
streamlit==1.25.0
streamlit-google-oauth @ git+https://github.com/hunkim/streamlit-google-oauth@5feb84d2b8298f0a512a017bf7ffa9672b449342
tenacity==8.4.1
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==4.3.1
urllib3==2.2.2
validators==0.28.3
watchdog==4.0.1
yarl==1.9.4
zipp==3.19.2
